requirements.txt
Flask
mysql-connector-python
redis
Flask-Mail
Pillow
python-dotenv
